requirements.txt
fastapi
uvicorn
python-multipart
sentence-transformers
transformers
torch
scikit-learn
PyPDF2
python-docx
numpy
pandas
python-dotenv
fastapi
uvicorn
python-multipart
sentence-transformers
transformers
torch
scikit-learn
PyPDF2
python-docx
numpy
pandas
python-dotenv
